Split db.py into a cohesive app/db package
Replace the 1278-line Database class with an app/db/ package: a
DatabaseEngine base (connection lifecycle, schema/migrations, shared
SQL helpers) composed with per-concern mixins (plans, deployment,
calls, transfers, termination, telemetry, webhooks, transcripts,
questions) behind the same flat Database facade and re-exports.
Deduplicate the mechanical patterns: _execute_cas for the repeated
rowcount==1 CAS idiom, _in_clause for hand-built IN placeholders,
_immediate_transaction for single-path transactional blocks (bespoke
rollback flows stay hand-rolled), one advisory-outcome serializer,
and a column allowlist on update_call mirroring set_flag_once.
