+++ app/db/calls.py
@@ -1,0 +1,163 @@
+from __future__ import annotations
+
+from collections.abc import Iterable
+from typing import Any
+
+from app.db.engine import _iso_now
+from app.models import TERMINAL_STATES, CallState
+
+# Schema columns of `calls` that current callers actually pass to update_call(). Anything
+# outside this set is rejected up front instead of failing later as a SQL error.
+_UPDATE_CALL_ALLOWED_COLUMNS = frozenset(
+    {
+        "state",
+        "conference_sid",
+        "twilio_ai_call_sid",
+        "twilio_callee_call_sid",
+        "twilio_owner_call_sid",
+        "openai_call_id",
+        "openai_accept_status",
+        "transcription_verified",
+        "semantic_vad_verified",
+        "callee_dialed",
+        "sideband_open",
+        "callee_joined",
+        "opening_sent",
+        "voicemail_sent",
+        "termination_claimed",
+        "termination_reason",
+        "transfer_outcome",
+        "advisory_outcome_json",
+        "interruption_observed",
+        "answered_at",
+        "last_event_at",
+    }
+)
+
+
+class CallsMixin:
+    async def get_call(self, call_id: str) -> dict[str, Any] | None:
+        return await self.fetch_one("SELECT * FROM calls WHERE call_id=?", (call_id,))
+
+    async def get_call_by_openai_id(self, openai_call_id: str) -> dict[str, Any] | None:
+        return await self.fetch_one("SELECT * FROM calls WHERE openai_call_id=?", (openai_call_id,))
+
+    async def get_call_by_twilio_sid(self, sid: str) -> dict[str, Any] | None:
+        return await self.fetch_one(
+            """SELECT * FROM calls
+               WHERE twilio_ai_call_sid=? OR twilio_callee_call_sid=?""",
+            (sid, sid),
+        )
+
+    async def list_calls(self, limit: int = 100) -> list[dict[str, Any]]:
+        return await self.fetch_all(
+            "SELECT * FROM calls ORDER BY created_at DESC LIMIT ?", (limit,)
+        )
+
+    async def list_nonterminal_calls(self) -> list[dict[str, Any]]:
+        placeholders, params = self._in_clause(state.value for state in TERMINAL_STATES)
+        return await self.fetch_all(
+            f"SELECT * FROM calls WHERE state NOT IN ({placeholders})",  # noqa: S608
+            params,
+        )
+
+    async def list_terminal_calls_needing_finalization(self) -> list[dict[str, Any]]:
+        placeholders, params = self._in_clause(state.value for state in TERMINAL_STATES)
+        return await self.fetch_all(
+            f"""SELECT calls.* FROM calls
+                LEFT JOIN call_results ON call_results.call_id = calls.call_id
+                WHERE calls.state IN ({placeholders})
+                  AND (
+                    call_results.call_id IS NULL
+                    OR json_extract(call_results.result_json, '$.finalization_status') =
+                       'telephony_only'
+                  )""",  # noqa: S608
+            params,
+        )
+
+    async def set_conference_cleanup_pending(self, call_id: str, pending: bool) -> None:
+        await self.execute(
+            "UPDATE calls SET conference_cleanup_pending=? WHERE call_id=?",
+            (1 if pending else 0, call_id),
+        )
+
+    async def list_conference_cleanup_pending(self) -> list[dict[str, Any]]:
+        return await self.fetch_all("SELECT * FROM calls WHERE conference_cleanup_pending=1")
+
+    async def touch_call(self, call_id: str) -> None:
+        await self.execute(
+            "UPDATE calls SET last_event_at=? WHERE call_id=?", (_iso_now(), call_id)
+        )
+
+    async def touch_calls(self, activity: Iterable[tuple[str, str]]) -> None:
+        """Persist latest observed activity for several calls in one durable transaction."""
+        placeholders, terminal_states = self._in_clause(
+            sorted(state.value for state in TERMINAL_STATES)
+        )
+        rows = [
+            (occurred_at, call_id, occurred_at, *terminal_states)
+            for call_id, occurred_at in activity
+        ]
+        if not rows:
+            return
+        async with self._immediate_transaction() as conn:
+            await conn.executemany(
+                f"""UPDATE calls SET last_event_at=?
+                    WHERE call_id=? AND last_event_at<?
+                      AND state NOT IN ({placeholders})""",  # noqa: S608
+                rows,
+            )
+
+    async def update_call(self, call_id: str, **values: Any) -> bool:
+        if not values:
+            return True
+        unknown = values.keys() - _UPDATE_CALL_ALLOWED_COLUMNS
+        if unknown:
+            raise ValueError(f"invalid call column(s): {', '.join(sorted(unknown))}")
+        values["last_event_at"] = _iso_now()
+        columns = ", ".join(f"{key}=?" for key in values)
+        params = [
+            self._serialize_advisory_outcome(value) if key == "advisory_outcome_json" else value
+            for key, value in values.items()
+        ]
+        params.append(call_id)
+        return await self._execute_cas(f"UPDATE calls SET {columns} WHERE call_id=?", params)
+
+    async def cas_state(self, call_id: str, expected: CallState, replacement: CallState) -> bool:
+        return await self._execute_cas(
+            """UPDATE calls SET state=?, last_event_at=?
+               WHERE call_id=? AND state=?""",
+            (replacement.value, _iso_now(), call_id, expected.value),
+        )
+
+    async def set_flag_once(self, call_id: str, flag: str) -> bool:
+        allowed = {
+            "sideband_open",
+            "callee_joined",
+            "callee_dialed",
+            "opening_sent",
+            "voicemail_sent",
+            "termination_claimed",
+        }
+        if flag not in allowed:
+            raise ValueError(f"invalid call flag: {flag}")
+        return await self._execute_cas(
+            f"UPDATE calls SET {flag}=1, last_event_at=? WHERE call_id=? AND {flag}=0",  # noqa: S608
+            (_iso_now(), call_id),
+        )
+
+    async def set_amd_once(self, call_id: str, answered_by: str, handling: str) -> bool:
+        return await self._execute_cas(
+            """UPDATE calls
+               SET amd_result=?, answered_by=?, answer_handling=?, last_event_at=?
+               WHERE call_id=? AND amd_result IS NULL""",
+            (answered_by, answered_by, handling, _iso_now(), call_id),
+        )
+
+    async def claim_opening_if_not_voicemail(self, call_id: str) -> bool:
+        """Atomically claim the opening unless AMD has already classified voicemail."""
+        return await self._execute_cas(
+            """UPDATE calls SET opening_sent=1, last_event_at=?
+               WHERE call_id=? AND opening_sent=0 AND answer_handling IS NOT ?""",
+            (_iso_now(), call_id, "voicemail"),
+        )
